requirements.txt
trafilatura==1.6.3
Levenshtein==0.23.0
google-cloud-vision==3.4.5
dotenv
boto3